+++ src/Backend/Renderer/GLES2/CanvasGLES2.cpp
@@ -109,7 +109,8 @@ namespace tgui
 
     void CanvasGLES2::setSize(const Layout2d& size)
     {
-        Vector2f newSize = size.getValue();
+        Widget::setSize(size);
+        const Vector2f newSize = getSize();
 
         if ((newSize.x > 0) && (newSize.y > 0))
         {
@@ -125,8 +126,6 @@ namespace tgui
 
             m_usedTextureSize = newTextureSize;
         }
-
-        Widget::setSize(size);
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////