+++ src/Backend/Renderer/SDL_Renderer/CanvasSDL.cpp
@@ -102,7 +102,8 @@ namespace tgui
 
     void CanvasSDL::setSize(const Layout2d& size)
     {
-        Vector2f newSize = size.getValue();
+        Widget::setSize(size);
+        const Vector2f newSize = getSize();
 
         if ((newSize.x > 0) && (newSize.y > 0))
         {
@@ -127,8 +128,6 @@ namespace tgui
 
             m_usedTextureSize = newTextureSize;
         }
-
-        Widget::setSize(size);
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////