+++ src/Backend/Renderer/SFML-Graphics/CanvasSFML.cpp
@@ -118,7 +118,8 @@ namespace tgui
 
     void CanvasSFML::setSize(const Layout2d& size)
     {
-        const Vector2f newSize = size.getValue();
+        Widget::setSize(size);
+        const Vector2f newSize = getSize();
 
         if ((newSize.x > 0) && (newSize.y > 0))
         {
@@ -132,8 +133,6 @@ namespace tgui
 
             m_usedTextureSize = newTextureSize;
         }
-
-        Widget::setSize(size);
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////