+++ src/Widgets/RichTextLabel.cpp
@@ -326,6 +326,7 @@ namespace tgui
         // Update the size of the label
         if (m_autoSize)
         {
+            m_autoLayout = AutoLayout::Manual;
             // NOLINTNEXTLINE(bugprone-parent-virtual-call)
             Widget::setSize({maxWidth + outline.getLeft() + outline.getRight() + 2*textOffset, requiredTextHeight + outline.getTop() + outline.getBottom()});
             m_bordersCached.updateParentSize(getSize());